server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
pytest==8.0.2
python-dotenv==1.0.1
fpdf2==2.8.1
gunicorn==21.2.0
